[fix] Image: style resolving
Cannot pass the result of StyleSheet.flatten to components, as it mixes
dynamic and static (compiled away) style information.

With the way Image is currently implemented, we have to override
'boxShadow' in all cases to avoid the 'shadow*' props being incorrectly
applied as a box-shadow. Once Image is implemented using createElement,
we can disable the 'boxShadow" generation just for Image.

Fix #2527